+++ tests/testthat/test-energy_balance.R
@@ -336,5 +336,20 @@ test_that("endogenize_losses() works correctly", {
     expect_true(Recca::balance_cols$losses_sector %in%
                   matsbyname::getcolnames_byname(endogenized$Y[[i]]))
   }
+
+  # Check that columns are NOT deleted when NOT requested
+  res1 <- pivoted_E |>
+    calc_intra_industry_balance() |>
+    endogenize_losses()
+  expect_true(Recca::balance_cols$intra_industry_balance_colname %in% names(res1))
+  expect_true("V_prime" %in% names(res1))
+  expect_true("Y_prime" %in% names(res1))
+  # Check that columns are deleted when requested
+  res2 <- pivoted_E |>
+    calc_intra_industry_balance() |>
+    endogenize_losses(replace_cols = TRUE)
+  expect_false(Recca::balance_cols$intra_industry_balance_colname %in% names(res2))
+  expect_false("V_prime" %in% names(res2))
+  expect_false("Y_prime" %in% names(res2))
 })